Starting mylar with --backup would cause an exception
I tried to start Mylar with the --backup switch, but it caused an unhandled exception.  So I updated the config_file to use the mylar.CONFIG_FILE, similar to the mylar.DB_FILE that is also backed up.

Result:  Starting Mylar with the --backup switch now copies the database and config files to the ./backup directory, as expected.  Renaming existing files if necessary.