log an error message when ComicVine has banned mylar's IP adddress for exceeding the rate limit.
This logs an error message when ComicVine returns HTML with the tag <title>Abnormal Traffic Detected....

The ban typically lasts a couple of days, and this may at least save the user some time in recognizing that a ban is the issue.

This is only a suggestion.  Feel free to close this if it's not a preferred solution!