getcomics: stop searching when first matching issue is found
GC is curated and so generally only has a single page per issue: this
commit assumes that we can therefore return the first match we find as
the only match.

Due to the use of a generator, the pages are fetched only when results
from the previous page are exhausted. By terminating iteration of the
generator early, therefore, we can potentially avoid fetching some
search result pages altogether: this reduces needless requests to GC,
and makes searches complete more quickly (as every additional fetch
incurs a 15s pause).