getcomics: handle literal-HTML descriptions which occasionally occur
I've run into issues on GC which have the short description HTML escaped
and so appearing as normal text: this means that it isn't queryable
using `.find`.  This commit detects that case and creates a new
`BeautifulSoup` object for the escaped HTML, and then searches it.

I've seen this fix searching for Batman (2016) #12, as it has such a
description.