getcomics: restore honouring of pack_priority configuration option
Instead of taking the _first_ result, we will now take the first result
which matches our `PACK_PRIORITY` setting.  (This will result in all
search pages being fetched if a result matching that setting does not
exist, but what else are we gonna do.)

I tested this with The Clone Conspiracy (2016): GC has both individual
books and a full 5-comic pack, and the pack is sorted before the
individual #1 book in the results.  Without this change, the pack is
always downloaded: with this change, `PACK_PRIORITY=True` continues to
download the pack, but `PACK_PRIORITY=False` continues searching and
finds the individual issue.