Making listlibrary handle multiple annuals
When a series had multiple annuals (and annual integration was on), listlibrary would only return one of them in the list, causing some to not show as in your library. By adding releasecomicid to the group, the DB call will return one row for each series/annual combo, and the resulting library will be complete.

Verified that this fixes both story arcs view and the search results view when looking for new series